update: PR review 반영
