--- src/main/java/com/goat/server/notification/dto/fcm/Data.java
@@ -1,9 +1,0 @@
-package com.goat.server.notification.dto.fcm;
-
-import lombok.Builder;
-
-@Builder
-public record Data(
-        String reviewId
-) {
-}
