--- src/main/java/com/goat/server/notification/dto/fcm/Message.java
@@ -1,11 +1,0 @@
-package com.goat.server.notification.dto.fcm;
-
-import lombok.Builder;
-
-@Builder
-public record Message(
-        Notification notification,
-        String token,
-        Data data
-) {
-}
