--- src/main/java/com/goat/server/notification/dto/fcm/Notification.java
@@ -1,11 +1,0 @@
-package com.goat.server.notification.dto.fcm;
-
-import lombok.Builder;
-
-@Builder
-public record Notification(
-        String title,
-        String body,
-        String image
-) {
-}
